requirements.txt
langchain
chromadb
openai
unstructured
fastapi
uvicorn
python-docx
PyMuPDF
pandas
geopandas
shapely
rasterio
folium
tqdm
langchain_community
llama-index-vector-stores-chroma
python-dotenv
python-multipart
numpy
transformers
sentence-transformers
accelerate
llama-index-embeddings-huggingface
sentencepiece
huggingface_hub
torch
llama-index-llms-huggingface
llama-index-core
